:hammer: refactor: Cache.hpp/cpp 추가 (전역변수 관리) 및 Client 코드 리팩토링
